factor out utils methods
